Fix group retention (#57)
* size -> recSize

* step in work

* work on count

* Text data source still have issues

* minus as filename part now is possible

* Right names

* Test for reading from txtsrc

* make cformat

* Fix G4 grammar - dead rule remove

* Work on UT

* Fix on test

* Work on test

* recreate test folders

* cmake format

* Work on ut

* style fix

* Using map for ordered sequence

* Work on transactional state file

* work on writeCount

* Remove webpage from derived branch (#54)

* Update .gitignore

This change cuts off webpage from master and only master and derived from master should cut this off

* Delete docs directory

This is webpage - this should not be in the master branch.

* Sync master to fix_group_retention (#55)

* Update .gitignore

This change cuts off webpage from master and only master and derived from master should cut this off

* Delete docs directory

This is webpage - this should not be in the master branch.

* bump conan deps

* Downgrade java for antlr

* CMake mess update

* Step down with boost v86->v85 - cause of errors. Waiting for v87 in conan

* Force CMake as minimum 3.10 due deprecation

* Remex octave script

* dsp script update

* dsp fix

* plotblock fix

* Cmake <3.10 deprecation

* Time

* Work on group

* seg-capacity retention

* Cmake 3.10 -> 3.20 req

* Work

* fix test

* testdb ignore

* name setter/getter intro

* Move logs to tmp/retractor

* vec on accessors in group

* segments on read

* write on retention segements

* segement remove support

* code guardians - asserts

* Developing tests, validating code

* Covering read and write

* Parser/lexer output remove from cformat

* Segments new approach

* work in progress

* work.

* candidate

* intro test

* remove comments

* Purge support

+++ src/include/rdb/descriptor.h
@@ -61,7 +61,7 @@ class Descriptor : public std::vector<rField> {
   std::string type(const std::string &name);
   int sizeFlat();
 
-  std::pair<int, int> retention();
+  std::pair<size_t, size_t> retention();
 
   std::pair<rdb::descFld, int> getMaxType();
